Update README.md
Everyone deserves to be understood, including signers! Sign language is often overlooked- and this application hopes to help rectify that. This is a real-time American Sign Language Gesture detector (may have alphabets added later), and its main goal is to detect the more common gestures in ASL, as those are often missed out on detectors. It uses mediapipe and TensorFlow to do so.